requirements.txt
supabase==2.5.1
requests==2.32.3
python-dotenv==1.0.1